debug config - trying to sort kernel panic -

+++ arch/arm64/include/asm/traps.h
@@ -24,9 +24,9 @@ struct undef_hook {
 	struct list_head node;
 	u32 instr_mask;
 	u32 instr_val;
-	u32 pstate_mask;
-	u32 pstate_val;
-	int (*fn)(struct pt_regs *regs, unsigned int instr);
+	u64 pstate_mask;
+	u64 pstate_val;
+	int (*fn)(struct pt_regs *regs, u32 instr);
 };
 
 void register_undef_hook(struct undef_hook *hook);
